refactor: split out container and content pages (#1119)
