+++ apps/watch/src/components/ShareDialog/ShareDialog.tsx
@@ -35,6 +35,13 @@ export function ShareDialog({
     setValue(newValue)
   }
 
+  const shareDescription =
+    video.description != null && video.description.length > 0
+      ? video.description[0].value
+      : video.snippet != null && video.snippet.length > 0
+      ? video.snippet[0].value
+      : ''
+
   const shareLink =
     routes != null
       ? `${
@@ -153,10 +160,7 @@ export function ShareDialog({
               {video.title[0].value}
             </Typography>
             <Typography>
-              {`${video.description[0].value
-                .split(' ')
-                .slice(0, 18)
-                .join(' ')}...`}
+              {`${shareDescription.split(' ').slice(0, 18).join(' ')}...`}
             </Typography>
           </Stack>
         </Stack>
