--- apps/watch/src/components/VideoContainer/index.ts
@@ -1,1 +1,0 @@
-export { VideoContainer } from './VideoContainer'
