+++ apps/watch/src/components/VideoContainerPage/index.ts
@@ -1,0 +1,1 @@
+export { VideoContainerPage } from './VideoContainerPage'
