+++ apps/watch/src/components/VideoContentPage/index.ts
@@ -1,0 +1,1 @@
+export { VideoContentPage } from './VideoContentPage'
